pyproject.toml
[project]
name = "mosaic"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "apscheduler>=3.11.1",
    "bcrypt>=3.2.0,<4.0.0",
    "claude-agent-sdk>=0.1.9",
    "click>=8.3.1",
    "email-threads",
    "fastapi>=0.124.4",
    "fastmcp>=2.13.3",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "prompt-toolkit>=3.0.52",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "pyzmq>=27.1.0",
    "rich>=14.2.0",
    "sqlmodel>=0.0.22",
    "tomli>=2.0.0",
    "tomli-w>=1.0.0",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mosaic"]

[tool.uv.sources]
email-threads = { git = "https://github.com/renjiyun06/email-threads.git" }

[project.scripts]
mosaic = "mosaic.refactor.cli.main:main"
mosaic_old = "mosaic.cli.main:mosaic"
